Move instance entity to data layer
